# swagger/doc.json
{
	"openapi": "3.0.1",
	"info": { "title": "API https://tvoe-dev.ru/", "version": "v1" },
	"servers": [{ "url": "http://localhost:4000" }],
...
	"paths": {
...
		"/collections/continueWatching": {
			"get": {
				"summary": "Подборка \"продолжить просмотр\"",
				"tags": ["Collections"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Кол-во пропускаемых элементов из начала выборки"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество документов"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"_id": {
														"type": "string",
														"description": "Id (mongoose.ObjectId) просмотра (документа коллекции MoviePageLog)"
													},
													"endTime": {
														"type": "number",
														"description": "Время окончания просмотра фильма в секундах"
													},
													"movieId": {
														"type": "string",
														"description": "Id фильма (mongoose.ObjectId)",
														"required": true
													},
													"videoId": {
														"type": "string",
														"description": "Id видео (mongoose.ObjectId)"
													},
													"updatedAt": {
														"type": "string",
														"description": "Дата обновления просмотра от пользователя"
													},
													"movie": {
														"type": "object",
														"properties": {
															"name": {
																"type": "string",
																"description": "Название фильма/сериала"
															},
															"alias": {
																"type": "string",
																"description": "alias фильма/сериала"
															},
															"categoryAlias": {
																"type": "string",
																"description": "Категория - serials или films"
															},
															"filmDuration": {
																"type": "number",
																"description": "Длительность фильма",
																"required": false
															},
															"poster": {
																"type": "object",
																"required": false,
																"properties": {
																	"src": {
																		"type": "string",
																		"description": "Путь к файлу изображению постера"
																	}
																}
															},
															"cover": {
																"type": "object",
																"required": false,
																"properties": {
																	"src": {
																		"type": "string",
																		"description": "Путь к файлу изображению обложки"
																	}
																}
															}
														}
													},
													"seriaInfo": {
														"type": "object",
														"required": false,
														"properties": {
															"season": {
																"type": "number",
																"description": "Номер сезона"
															},
															"episode": {
																"type": "number",
																"description": "Номер серии"
															},
															"seriaDuration": {
																"type": "number",
																"description": "Длительность серии"
															}
														}
													}
												}
											}
										}
									},
									"example": {
										"value": {
											"totalSize": 29,
											"items": [
												{
													"_id": "64e8801f6ffabb5ba3d15704",
													"endTime": 4.710858,
													"movieId": "64d7871e047dca3ed1b82899",
													"videoId": "64d787a8047dca3ed1b82cc2",
													"updatedAt": "2023-08-25T10:19:11.336Z",
													"movie": {
														"name": "Сериал тест hls",
														"alias": "serial-test-hls",
														"categoryAlias": "serials"
													},
													"seriaInfo": {
														"season": 2,
														"episode": 1,
														"seriaDuration": 4170
													}
												},
												{
													"_id": "64e87d576ffabb5ba3d14f03",
													"endTime": 3.636435,
													"movieId": "64d7871e047dca3ed1b82899",
													"videoId": "64d787a8047dca3ed1b82cc1",
													"updatedAt": "2023-08-25T10:07:19.935Z",
													"movie": {
														"name": "Сериал тест hls",
														"alias": "serial-test-hls",
														"categoryAlias": "serials"
													},
													"seriaInfo": {
														"season": 1,
														"episode": 2,
														"seriaDuration": 4170
													}
												},
												{
													"_id": "64e72f736ffabb5ba3ce617d",
													"endTime": 8.443566,
													"movieId": "64b64f423d8baf6888235f9c",
													"videoId": "64e72ed2ea4ca856d3ab6fc6",
													"updatedAt": "2023-08-24T10:22:43.628Z",
													"movie": {
														"name": "Тестовый фильм название",
														"alias": "just-test",
														"categoryAlias": "films",
														"filmDuration": 4170
													},
													"seriaInfo": null
												}
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/collections/possibleYouLike": {
			"get": {
				"summary": "Подборка - \"возможно вам понравится\"",
				"tags": ["Collections"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Кол-во пропускаемых элементов из начала выборки"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество документов"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"name": {
														"type": "string",
														"description": "Название фильма/сериала"
													},
													"_id": {
														"type": "string",
														"description": "Id (mongoose.ObjectId) просмотра (документа коллекции Movies)"
													},
													"duration": {
														"type": "number",
														"description": "Длительность"
													},
													"url": {
														"type": "string",
														"description": "Url к фильму"
													},
													"poster": {
														"type": "object",
														"required": false,
														"properties": {
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению обложки"
															}
														}
													},
													"cover": {
														"type": "object",
														"required": false,
														"properties": {
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению обложки"
															}
														}
													}
												}
											}
										}
									},
									"example": {
										"value": {
											"totalSize": 389,
											"items": [
												{
													"_id": "63b9a9934b33357ef7068057",
													"name": "Как я стал супергероем",
													"poster": {
														"_id": "63ba2c404b33357ef70698bb",
														"src": "/images/63ba2c404b33357ef70698bc.jpg"
													},
													"cover": {
														"_id": "63ba2cdc4b33357ef70698d4",
														"src": "/images/63ba2cdc4b33357ef70698d5.jpg"
													},
													"duration": 6063.791667,
													"url": "/p/kak-ya-stal-supergeroem"
												},
												{
													"_id": "63b866da4b33357ef7063e22",
													"name": "Обитатели холмов",
													"poster": {
														"_id": "63c3d0c61c98adb3690e02d8",
														"src": "/images/63c3d0c61c98adb3690e02d9.jpg"
													},
													"cover": {
														"_id": "63c3d0c71c98adb3690e02f1",
														"src": "/images/63c3d0c71c98adb3690e02f2.jpg"
													},
													"duration": 12246.25,
													"url": "/p/obitateli-holmov"
												},
												{
													"_id": "63dcfff61311449b80d87352",
													"name": "Академия Амбрелла",
													"cover": {
														"_id": "63e6c7025c8c77c4bcc60b48",
														"src": "/images/63e6c7025c8c77c4bcc60b49.jpg"
													},
													"poster": {
														"_id": "63e6c6ee5c8c77c4bcc60b2b",
														"src": "/images/63e6c6ee5c8c77c4bcc60b2c.jpg"
													},
													"duration": 61911.583333,
													"url": "/p/akademiya-ambrella"
												}
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}